examples/android/benchmarks: Add analysis for sched dvfs util
In schedutil governor it needs to accumulate the util from RT, CFS
scheduling classes and finally calculate the total utilization. We can
analyse these sched dvfs utilization values to check if the workloads
have been spreads out to CPUs within the same cluster so this can
benefit both the scheduling latency and power with lower OPPs.

This commit is to add analysis for sched dvfs util for mean value for
all CPUs and calculate the standard deviation for every kind util value.

Signed-off-by: Leo Yan <leo.yan@linaro.org>